Fixed error in Admin -> Configuration after added custom scripts #5113

+++ packages/Webkul/Core/src/Repositories/CoreConfigRepository.php
@@ -91,7 +91,9 @@ class CoreConfigRepository extends Repository
                     ]);
                 } else {
                     foreach ($coreConfigValue as $coreConfig) {
-                        Storage::delete($coreConfig['value']);
+                        if (request()->hasFile($fieldName)) {
+                            Storage::delete($coreConfig['value']);
+                        }
 
                         if (isset($value['delete'])) {
                             $this->model->destroy($coreConfig['id']);
